http api: more correct types, less build warnings
build using gcc (4.8.4) now completes 100% quiet (clang not yet)

Also fixes DBG to not generate unused expression/value warnings when not used.

+++ httphandler.c
@@ -101,7 +101,7 @@ int uhStats(UrlHandlerParam* param)
 	mwWriteXmlString(&p, &bufsize, 0, "</ServerStats>");
 
 	//return data to server
-	param->dataBytes=(int)p-(int)(param->pucBuffer);
+	param->dataBytes=(int)(p - param->pucBuffer);
 	param->fileType=HTTPFILETYPE_XML;
 	return ret;
 }
